+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-06-16 17:10:30
-// Commit: feat(network): add bi-directional event bus handler for network
+// Updated at: 2026-06-16 20:51:38
+// Commit: docs(network): update readme and API contracts for network
 
 class networkComponentState {
-  final String commitRef = '2036';
-  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
-  final DateTime timestamp = DateTime.parse('2026-06-16T17:10:30+05:00');
+  final String commitRef = '2048';
+  final String lastMessage = 'docs(network): update readme and API contracts for network';
+  final DateTime timestamp = DateTime.parse('2026-06-16T20:51:38+05:00');
 }